Create  Removing duplicate elements
The way to think about this problem is you have two pointers, under a while loop. the first pointer is your judgement pointer you use this to decide how you want to place your elements.
You move the second pointer indefinitely but the first pointer you move when both elements are different, and when they are different you change the value of the element after the one being repersented by the first pointer so j+1. Then you move your first pointer to that value, this insures you dont get any duplicates and correctly sort the array.