Create Pascals Triangle
The Concept behind pascals triangle is that you generate a set number of rows represented by the variable numRows, each row starts and ends with 1. Each new row depends on the size and elements of the previous row. we assign the variable prev the previous row in triangle so if we are on iteration 2 the row before that would be [1,1]. Each row starts with one, then we use a for loop around the prev loop starting at one and then going through the  length of the loop. in this case the len(prev) would only add one more character and then we finish each row of by adding one.